removed extended process event
