+++ activiti-cloud-api/activiti-cloud-api-events/src/test/java/org/activiti/cloud/api/events/CloudRuntimeEventTypeTest.java
@@ -31,7 +31,6 @@ import org.activiti.api.process.model.events.ProcessDefinitionEvent;
 import org.activiti.api.process.model.events.ProcessRuntimeEvent;
 import org.activiti.api.process.model.events.SequenceFlowEvent;
 import org.activiti.api.process.model.events.ApplicationEvent;
-import org.activiti.cloud.api.process.model.events.ExtendedCloudProcessRuntimeEvent;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -51,7 +50,6 @@ public class CloudRuntimeEventTypeTest {
         assertThat(enumValues).containsAll(this.getEnumValuesAsStringList(BPMNErrorReceivedEvent.ErrorEvents.class));
         assertThat(enumValues).containsAll(this.getEnumValuesAsStringList(BPMNSignalEvent.SignalEvents.class));
         assertThat(enumValues).containsAll(this.getEnumValuesAsStringList(ProcessDefinitionEvent.ProcessDefinitionEvents.class));
-        assertThat(enumValues).containsAll(this.getEnumValuesAsStringList(ExtendedCloudProcessRuntimeEvent.ExtendedCloudProcessRuntimeEvents.class));
         assertThat(enumValues).containsAll(this.getEnumValuesAsStringList(IntegrationEvent.IntegrationEvents.class));
         assertThat(enumValues).containsAll(this.getEnumValuesAsStringList(BPMNTimerEvent.TimerEvents.class));
         assertThat(enumValues).containsAll(this.getEnumValuesAsStringList(BPMNMessageEvent.MessageEvents.class));
