--- activiti-cloud-audit-service/activiti-cloud-services-audit-jpa/src/main/java/org/activiti/cloud/services/audit/jpa/events/ExtendedProcessAuditEventEntity.java
@@ -1,32 +1,0 @@
-/*
- * Copyright 2017-2020 Alfresco Software, Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.activiti.cloud.services.audit.jpa.events;
-
-import javax.persistence.MappedSuperclass;
-import org.activiti.cloud.api.process.model.events.ExtendedCloudProcessRuntimeEvent;
-
-@MappedSuperclass
-public abstract class ExtendedProcessAuditEventEntity extends ProcessAuditEventEntity {
-
-    public ExtendedProcessAuditEventEntity() {
-    }
-
-    public ExtendedProcessAuditEventEntity(ExtendedCloudProcessRuntimeEvent cloudEvent) {
-        super(cloudEvent);
-        setProcessInstance(cloudEvent.getEntity());
-    }
-
-}
