+++ activiti-cloud-audit-service/activiti-cloud-services-audit-jpa/src/main/java/org/activiti/cloud/services/audit/jpa/events/ProcessDeletedAuditEventEntity.java
@@ -23,7 +23,7 @@ import org.hibernate.annotations.DynamicInsert;
 @Entity(name = ProcessDeletedAuditEventEntity.PROCESS_DELETED_EVENT)
 @DiscriminatorValue(value = ProcessDeletedAuditEventEntity.PROCESS_DELETED_EVENT)
 @DynamicInsert
-public class ProcessDeletedAuditEventEntity extends ExtendedProcessAuditEventEntity {
+public class ProcessDeletedAuditEventEntity extends ProcessAuditEventEntity {
 
     protected static final String PROCESS_DELETED_EVENT = "ProcessDeletedEvent";
