+++ activiti-cloud-query-service/activiti-cloud-services-query/activiti-cloud-services-query-rest/src/test/java/org/activiti/cloud/services/query/events/handlers/QueryEventHandlerContextIT.java
@@ -31,7 +31,6 @@ import org.activiti.api.process.model.events.SequenceFlowEvent;
 import org.activiti.api.task.model.events.TaskCandidateGroupEvent;
 import org.activiti.api.task.model.events.TaskCandidateUserEvent;
 import org.activiti.api.task.model.events.TaskRuntimeEvent;
-import org.activiti.cloud.api.process.model.events.ExtendedCloudProcessRuntimeEvent.ExtendedCloudProcessRuntimeEvents;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -57,7 +56,7 @@ public class QueryEventHandlerContextIT {
                 ProcessRuntimeEvent.ProcessEvents.PROCESS_RESUMED.name(),
                 ProcessRuntimeEvent.ProcessEvents.PROCESS_CANCELLED.name(),
                 ProcessRuntimeEvent.ProcessEvents.PROCESS_UPDATED.name(),
-                ExtendedCloudProcessRuntimeEvents.PROCESS_DELETED.name(),
+                ProcessRuntimeEvent.ProcessEvents.PROCESS_DELETED.name(),
                 TaskRuntimeEvent.TaskEvents.TASK_CREATED.name(),
                 TaskRuntimeEvent.TaskEvents.TASK_ASSIGNED.name(),
                 TaskRuntimeEvent.TaskEvents.TASK_SUSPENDED.name(),
@@ -94,7 +93,7 @@ public class QueryEventHandlerContextIT {
         assertThat(handlers.get(ProcessRuntimeEvent.ProcessEvents.PROCESS_RESUMED.name())).isInstanceOf(ProcessResumedEventHandler.class);
         assertThat(handlers.get(ProcessRuntimeEvent.ProcessEvents.PROCESS_CANCELLED.name())).isInstanceOf(ProcessCancelledEventHandler.class);
         assertThat(handlers.get(ProcessRuntimeEvent.ProcessEvents.PROCESS_UPDATED.name())).isInstanceOf(ProcessUpdatedEventHandler.class);
-        assertThat(handlers.get(ExtendedCloudProcessRuntimeEvents.PROCESS_DELETED.name())).isInstanceOf(ProcessDeletedEventHandler.class);
+        assertThat(handlers.get(ProcessRuntimeEvent.ProcessEvents.PROCESS_DELETED.name())).isInstanceOf(ProcessDeletedEventHandler.class);
         assertThat(handlers.get(TaskRuntimeEvent.TaskEvents.TASK_CREATED.name())).isInstanceOf(TaskCreatedEventHandler.class);
         assertThat(handlers.get(TaskRuntimeEvent.TaskEvents.TASK_ASSIGNED.name())).isInstanceOf(TaskAssignedEventHandler.class);
         assertThat(handlers.get(TaskRuntimeEvent.TaskEvents.TASK_SUSPENDED.name())).isInstanceOf(TaskSuspendedEventHandler.class);
